feat: Add manual trigger and TestPyPI publishing option
This commit introduces a manual trigger for the "Test PyPI Release" workflow, allowing for on-demand testing of the PyPI release process.

A new input `test_pypi` has been added to the workflow_dispatch configuration. This boolean input, when set to true, enables publishing to TestPyPI. This feature is disabled by default and can be enabled manually when triggering the workflow.